+++ lib/src/lib.rs
@@ -122,6 +122,7 @@ pub mod view;
 pub mod working_copy;
 pub mod workspace;
 pub mod workspace_store;
+pub mod workspace_util;
 
 #[cfg(test)]
 mod tests {
